renamed config package to controller
